backend/requirements.txt
Flask
flask-cors
pymongo[srv]
python-dotenv
cohere